requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==0.1.22
openai==1.45.0
chromadb==0.5.5
faiss-cpu==1.8.0.post1
tiktoken==0.7.0
pypdf==4.3.1
unstructured==0.15.9
python-multipart==0.0.9
streamlit==1.37.1
